Add Register component and enhance Login UI
Introduces a new Register.vue component with a multi-step registration form, validation, and improved user experience. The Login.vue component is redesigned with a modern UI, validation, and social login options. AppHeader scroll logic is improved for navigation, and minor style adjustments are made to LandingPage.vue. The login page now uses the Login component for consistency.
